Delete C8_1 rule and replace with C8_2
