+++ dataactvalidator/config/sqlrules/sqlRules.csv
@@ -84,8 +84,7 @@ C6,GrossOutlaysUndeliveredOrdersPrepaidTotal (FYB or CPE) = USSGL(4802 + 4882). 
 C6,GrossOutlaysUndeliveredOrdersPrepaidTotal (FYB or CPE) = USSGL(4802 + 4882). This applies to the award level.,GrossOutlaysUndeliveredOrdersPrepaidTotalFYB = USSGL4802. This applies to the award level.,n,award_financial,warning,,c6_award_financial_2
 C7,GrossOutlaysDeliveredOrdersPaidTotal (FYB or CPE) = USSGL(4902 + 4908 + 4982). This applies to the award level.,GrossOutlaysDeliveredOrdersPaidTotalCPE = USSGL(4902 + 4908 + 4982). This applies to the award level.,n,award_financial,warning,,c7_award_financial_1
 C7,GrossOutlaysDeliveredOrdersPaidTotal (FYB or CPE) = USSGL(4902 + 4908 + 4982). This applies to the award level.,GrossOutlaysDeliveredOrdersPaidTotalFYB = USSGL4908. This applies to the award level.,n,award_financial,warning,,c7_award_financial_2
-C8,Invalid AllocationTransferAgency in File C (award financial).,Invalid AllocationTransferAgency in File C (award financial).,n,award_financial,warning,,c8_award_financial_1
-C8,"Unique FAIN/URI from file C exists in file D2. FAIN may be null for aggregated records.  URI may be null for non-aggregated records.",Unique FAIN/URI from file C exists in file D2. FAIN may be null for aggregated records.  URI may be null for non-aggregated records.,y,award_financial,warning,award,c8_award_financial_2
+C8,"Unique FAIN/URI from file C exists in file D2. FAIN may be null for aggregated records.  URI may be null for non-aggregated records.",Unique FAIN/URI from file C exists in file D2. FAIN may be null for aggregated records.  URI may be null for non-aggregated records.,y,award_financial,warning,award,c8_award_financial_1
 C9,"Unique FAIN/URI from file D2 exists in file C, except D2 records where FederalActionObligation and OriginalLoanSubsidyCost = 0. FAIN may be null for aggregated records.  URI may be null for non-aggregated records.","Unique FAIN/URI from file D2 exists in file C, except D2 records where FederalActionObligation and OriginalLoanSubsidyCost = 0. FAIN may be null for aggregated records.  URI may be null for non-aggregated records.",y,award,warning,award_financial,c9_award_financial
 C11,"Unique PIID, ParentAwardId from file C exists in file D1 during the same reporting period.","Unique PIID, ParentAwardId from file C exists in file D1 during the same reporting period.",y,award_financial,warning,award_procurement,c11_cross_file
 C12,"Unique PIID, ParentAwardId from file D1 exists in file C during the same reporting period, except D1 records with zero FederalActionObligation","Unique PIID, ParentAwardId from file D1 exists in file C during the same reporting period, except D1 records with zero FederalActionObligation",y,award_procurement,warning,award_financial,c12_cross_file
